@@ -23,8 +23,8 @@
     "express": "^4.14.0",
     "nodemon": "^2.0.4",
     "pg": "^8.7.3",
-    "react": "^16.3.2",
-    "react-dom": "^16.3.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-redux": "^5.0.7",
     "react-router-dom": "^4.2.2",
     "redux": "^4.0.0",
